+++ src/engine/Graphics/Context/screen-pass-painter/screen-pass-painter.ts
@@ -46,8 +46,14 @@ export class ScreenPassPainter {
 
   renderWithPostProcessor(postprocessor: PostProcessor): void {
     const gl = this._gl;
-    postprocessor.getShader().use();
+    const shader = postprocessor.getShader();
+    shader.use();
     postprocessor.getLayout().use();
+    gl.activeTexture(gl.TEXTURE0);
+    shader.trySetUniformInt('u_image', 0);
+    if (postprocessor.onDraw) {
+      postprocessor.onDraw();
+    }
     gl.drawArrays(gl.TRIANGLES, 0, 6);
   }