+++ src/engine/Graphics/PostProcessor/PostProcessor.ts
@@ -28,4 +28,10 @@ export interface PostProcessor {
    * @param elapsed
    */
   onUpdate?(elapsed: number): void;
+
+  /**
+   * Use the onDraw hook to upload any textures or command that need to run right before draw
+   * @param elapsed
+   */
+  onDraw?(): void;
 }